# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Randevu listesini getir",
                "operationId": "812c4f15927f3d0e3e959538a6f1104b",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Seçilen tarih (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "view",
                        "in": "query",
                        "description": "Listeleme tipi (daily veya weekly)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Randevular listelendi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Yeni randevu oluştur",
                "operationId": "2b481df28ec53d2afb2ec7c9a3a77846",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "employee_id",
                                    "start_datetime"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "employee_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "start_datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-17 10:30:00"
                                    },
                                    "end_datetime": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-12-17 11:00:00"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "Ek not"
                                    },
                                    "service": {
                                        "type": "string",
                                        "example": "Kuaför"
                                    },
                                    "cost": {
                                        "type": "number",
                                        "example": 120.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Randevu başarıyla eklendi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/appointments/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Belirli randevuyu getir",
                "operationId": "cca217dc71eae262de1fe09f314a4948",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Randevu bulundu"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Randevuyu güncelle",
                "operationId": "56cd9e193d590ed31b28bf8fe7162e70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "employee_id": {
                                        "type": "integer"
                                    },
                                    "start_datetime": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end_datetime": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "service": {
                                        "type": "string"
                                    },
                                    "cost": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Randevu güncellendi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/appointments/customer/{customerId}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Müşteriye ait randevuları getir",
                "operationId": "0e595241f1ca6830e3ce83ac5a662879",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Randevular listelendi"
                    },
                    "404": {
                        "description": "Randevu bulunamadı"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}